--- test/ads-lost-err_test.go
@@ -1,11 +1,0 @@
-package test
-
-import (
-	"testing"
-
-	"github.com/mgechev/revive/rule"
-)
-
-func TestADSLostErr(t *testing.T) {
-	testRule(t, "ads-lost-err", &rule.ADSLostErrRule{})
-}
